+++ Runtime/ColorVariable.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace UnderLogic.Variables
+{
+    [CreateAssetMenu(menuName = "Variables/Color Variable")]
+    public class ColorVariable : RuntimeVariable<Color> { }
+}