+++ Runtime/References/ColorReference.cs
@@ -1,0 +1,8 @@
+using System;
+using UnityEngine;
+
+namespace UnderLogic.Variables.References
+{
+    [Serializable]
+    public class ColorReference : RuntimeReference<Color, ColorVariable> { }
+}