+++ Runtime/Watchers/ColorVariableWatcher.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+namespace UnderLogic.Variables.Watchers
+{
+    public class ColorVariableWatcher : RuntimeVariableWatcher<Color, ColorVariable> { }
+}