+++ CHANGELOG.md
@@ -4,12 +4,21 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2022-12-03
+
+### Added
+
+- `ColorVariable` variant
+- `ColorVariableWatcher` component
+- `ColorReference` variant
+- Custom editors for new types
+
 ## [0.4.2] - 2022-12-03
 
 ### Added
 
-- Added `isReadOnly` and `IsReadOnly` property to `RuntimeVariable` base class
-- Added `MakeReadOnly()` method to `RuntimeVariable` base class
+- `isReadOnly` and `IsReadOnly` property to `RuntimeVariable` base class
+- `MakeReadOnly()` method to `RuntimeVariable` base class
 
 ## [0.4.1] - 2022-12-02